hPa value int

+++ softwarePico/libs/sensors.py
@@ -108,7 +108,7 @@ def measure(time_DTF):
             measures['temperature'] += th_s.temperature
             measures['humidity'] += th_s.relative_humidity
             pressure = bm_b.pressure # Pascal
-            p_hPa = pressure/100
+            p_hPa = int(pressure/100)
             measures['barometric pressure'] += p_hPa
             if use_aux_sensor:
                 pm_0_data = pm_p.measure()['mass_density']
